pyproject.toml
[project]
name = "ai-tf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.0",
    "langchain-chroma>=0.2.4",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.1",
    "python-dotenv>=1.1.1",
    "fastapi[standard]>=0.116.0",
    "pydantic>=2.0.0",
    "google-generativeai>=0.8.5",
]

[dependency-groups]
dev = [
    "fastapi>=0.116.0",
]
